chore: update .gitignore, add example configuration, and implement profiler
- Updated .gitignore to include  'screeps.json'.
- Added 'example.screeps.json' for sample Screeps server configuration.
- Implemented a new profiler module in 'lib/helpers/profiler.rs' to track performance metrics.
- Updated 'lib/helpers/mod.rs' to include the new profiler module.
- Refactored profiling logic from 'lib/utils.rs' into the new profiler module for better organization and maintainability.
