+++ lib/helpers/mod.rs
@@ -1,1 +1,2 @@
 pub mod cost_matrix;
+pub mod profiler;
